Changes components locations and connect BookList to the store
