+++ src/components/Book.js
@@ -2,11 +2,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const Book = ({ book }) => (
-  <>
-    <td>{book.id}</td>
-    <td>{book.title}</td>
-    <td>{book.category}</td>
-  </>
+  <table className="table">
+    <tr>
+      <td>{book.id}</td>
+      <td>{book.title}</td>
+      <td>{book.category}</td>
+    </tr>
+  </table>
 );
 
 Book.propTypes = {
